smart_legal_tool/requirements.txt
openai
langchain
langchain-openai
langchain-huggingface
python-dotenv
docarray
transformers
torch

